+++ src/themes/theme.tsx
@@ -1,0 +1,20 @@
+import { createTheme } from '@mui/material';
+
+const theme = createTheme({
+  palette: {
+    primary: {
+      main: '#16423C'
+    },
+    secondary: {
+      main: '#6A9C89'
+    },
+    background: {
+      paper: '#E9EFEC'
+    },
+    text: {
+      primary: '#16423C'
+    }
+  }
+});
+
+export default theme;
